HBASE-25842 move regionserver http-related code into o.a.h.h.regionserver.http
